pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nightwire"
version = "3.0.2"
description = "A Signal messaging bot with Claude AI integration for intelligent conversations and autonomous development"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "hackingdave"},
]
dependencies = [
    "aiohttp>=3.8",
    "pydantic>=2.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0",
    "structlog>=23.0",
    "sentence-transformers>=2.2.0",
    "sqlite-vec>=0.1.1",
    "psutil>=5.9",
]

[project.urls]
Repository = "https://github.com/offsecginger/nightwire"
"Bug Tracker" = "https://github.com/offsecginger/nightwire/issues"

[project.scripts]
nightwire = "nightwire.main:run"

[project.optional-dependencies]
sdk = [
    "anthropic>=0.77.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
    "psutil>=5.9",
    "mypy>=1.0",
    "bandit>=1.7",
    "safety>=2.3",
    "anthropic>=0.77.0",
]
autonomous = [
    "pytest-json-report>=1.0",
]

[tool.setuptools.packages.find]
include = ["nightwire*"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.bandit]
exclude_dirs = ["tests", "venv"]
skips = ["B101"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
